🚚 refactor: move api endpoints for better organization
