+++ ui/pages/dashboard/[owner]/[repo]/release/zenodo.vue
@@ -1,0 +1,57 @@
+<script setup lang="ts">
+import { useBreadcrumbsStore } from "@/stores/breadcrumbs";
+
+const breadcrumbsStore = useBreadcrumbsStore();
+
+breadcrumbsStore.showBreadcrumbs();
+breadcrumbsStore.setFeature({
+  id: "release-zenodo",
+  name: "Zenodo",
+  icon: "simple-icons:zenodo",
+});
+
+// const { data, error } = await useFetch(
+//   `/api/dashboard/${owner}/${repo}/release/zenodo`,
+//   {
+//     headers: useRequestHeaders(["cookie"]),
+//   },
+// );
+</script>
+
+<template>
+  <main class="mx-auto max-w-screen-xl px-8 pb-8 pt-4">
+    <n-flex vertical>
+      <h1>Zenodo Release</h1>
+
+      <p>
+        Create a release of your repository on Zenodo. We will also create a
+        Github release and update all referencing DOIs to point to the Zenodo
+        release.
+      </p>
+    </n-flex>
+
+    <CardPlaceholder placeholder="Confirm license" />
+
+    <CardPlaceholder placeholder="Confirm code metadata" />
+
+    <n-divider />
+
+    <h2 class="pb-6">Login to Zenodo</h2>
+
+    <CardPlaceholder placeholder="Login to Zenodo" />
+
+    <CardPlaceholder placeholder="Select Zenodo deposition or create new" />
+
+    <CardPlaceholder placeholder="?Select Zenodo community?" />
+
+    <CardPlaceholder placeholder="Add zenodo required metadata" />
+
+    <CardPlaceholder
+      placeholder="Create draft github release (add additional executables and update release notes)"
+    />
+
+    <CardPlaceholder placeholder="Confirm github release" />
+
+    <CardPlaceholder placeholder="Publish Zenodo release" />
+  </main>
+</template>
